+++ src/components/boards/BestPostList.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import BestPost from "./BestPost";
+import BestPost from "@/components/boards/BestPost";
 import useDeviceSize from "@/hooks/useDeviceSize";
 
 const BEST_POST_LIMIT: { [key in DeviceSizes]: number } = {
